+++ app/src/components/common/index.ts
@@ -14,3 +14,4 @@ export * from './Input';
 /* etc */
 export * from './Overlay';
 export * from './StyledInfoTable';
+export * from './Scroll';
