+++ app/src/pages/EvalLogSearchPage/index.ts
@@ -1,0 +1,1 @@
+export * from './EvalLogSearchPage';
