--- app/src/pages/EvaluationLogSearchPage/index.ts
@@ -1,1 +1,0 @@
-export * from './EvaluationLogSearchPage';
