+++ app/src/utils/formatters/index.ts
@@ -2,3 +2,4 @@ export * from './dDayFormatter';
 export * from './dateFormatter';
 export * from './numberFormatter';
 export * from './numberWithUnitFormatter';
+export * from './snakeCaseFormatter';
