+++ app/src/utils/formatters/snakeCaseFormatter.ts
@@ -1,0 +1,3 @@
+export const snakeCaseFormatter = (str: string) => {
+  return str.split(' ').join('_').toLowerCase();
+};
